added readonly functionality, refactored directory and naming
