+++ packages/plugins-image/src/components/drag-drop-archive
