+++ stories/plugins/image.stories.js
@@ -9,45 +9,50 @@ import initialState from "../states/image.json";
 import Editor from "../support/components/editor";
 
 const customOptions = {
-  maxFileSize: 1000000,
-  style: {
-    backgroundColor: "grey",
-    height: "15rem"
-  },
-  renderToolbar: ({ hovering, loading }, tools) => {
-    return (
-      <div>
-        {tools.map(tool => <button onClick={tool.action}>{tool.name}</button>)}
-      </div>
-    );
-  },
-  renderSelect: (selectFile, deleteImage) => {
-    return (
-      <div>
-        <button onClick={selectFile}>Select File</button>
-        <button onClick={deleteImage}>Delete</button>
-      </div>
-    );
-  },
-  renderErrors: error => {
-    return <div style={{ color: "red" }}>{error}</div>;
-  },
-  uploadImage: (file, updateSrc) => {
-    let data = new FormData();
-    data.append("file", file);
-
-    fetch("http://localhost:4000/api/upload", {
-      method: "POST",
-      body: data,
-      "Content-Type": file.type
-    })
-      .then(resp => {
-        return resp.text();
-      })
-      .then(newUrl => {
-        updateSrc(newUrl);
-      });
-  }
+  // maxFileSize: 1000000,
+  // style: {
+  //   backgroundColor: "whitesmoke",
+  //   height: "15rem"
+  // },
+  // renderToolbar: ({ hovering, loading }, tools) => {
+  //   return (
+  //     hovering &&
+  //     <div>
+  //       {tools.map(tool => <button onClick={tool.action}>{tool.name}</button>)}
+  //     </div>
+  //   );
+  // },
+  // renderSelect: (selectFile, deleteImage) => {
+  //   return (
+  //     <div>
+  //       <button onClick={selectFile}>Select File</button>
+  //       <button onClick={deleteImage}>Delete</button>
+  //     </div>
+  //   );
+  // },
+  // renderError: error => {
+  //   return <div style={{ color: "red" }}>{error}</div>;
+  // },
+  // uploadImage: (file, updateSrc, updateErrors) => {
+  //   let data = new FormData();
+  //   data.append("file", file);
+  //   fetch("http://localhost:4000/api/upload", {
+  //     method: "POST",
+  //     body: data,
+  //     "Content-Type": file.type
+  //   })
+  //   .then(resp => {
+  //     return resp.text();
+  //   })
+  //   .then(newUrl => {
+  //     updateSrc(newUrl);
+  //   })
+  //   .catch((e) => {
+  //     updateSrc();
+  //     updateErrors('Failed to upload file to server');
+  //   })
+  //   ;
+  // }
 };
 
 const pluginOpts = [
